Polishing the code
Tweaking classes, introducing API, spliting to diffrent modules.
Adding tests.
