+++ .travis.yml
@@ -1,8 +1,8 @@
 language: java
 sudo: false
 dist: trusty
-install: mvn install -DskipTests=true -Dmaven.javadoc.skip=true -B -V
-script: mvn clean verify --fail-at-end -U -V
+install: mvn clean package --batch-mode -DskipTests=true -Dmaven.javadoc.skip=true -B -V
+script: mvn verify --fail-at-end -U -V
 notifications:
   email:
     on_failure: change
