+++ sampler-api/src/main/java/pl/wavesoftware/sampler/api/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * @author <a href="mailto:krzysztof.suszynski@wavesoftware.pl">Krzysztof Suszynski</a>
+ * @since 0.1.0
+ */
+
+@ParametersAreNonnullByDefault
+@ReturnTypesAreNonnullByDefault
+package pl.wavesoftware.sampler.api;
+
+import pl.wavesoftware.eid.api.ReturnTypesAreNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
