+++ sampler-example-model/src/main/java/pl/wavesoftware/sampler/example/sample/group/RemoteDesktopSampler.java
@@ -1,0 +1,14 @@
+package pl.wavesoftware.sampler.example.sample.group;
+
+import pl.wavesoftware.sampler.example.model.Group;
+
+/**
+ * @author <a href="mailto:krzysztof.suszynski@wavesoftware.pl">Krzysztof Suszynski</a>
+ * @since 0.1.0
+ */
+final class RemoteDesktopSampler implements RemoteDesktop {
+  @Override
+  public Group create() {
+    return new Group("remote-desktop");
+  }
+}
