+++ sampler-example-model/src/main/java/pl/wavesoftware/sampler/example/sample/user/JohnDoe.java
@@ -1,0 +1,11 @@
+package pl.wavesoftware.sampler.example.sample.user;
+
+import pl.wavesoftware.sampler.api.Sampler;
+import pl.wavesoftware.sampler.example.model.User;
+
+/**
+ * @author <a href="mailto:krzysztof.suszynski@wavesoftware.pl">Krzysztof Suszynski</a>
+ * @since 0.1.0
+ */
+public interface JohnDoe extends Sampler<User> {
+}
