+++ sampler-spring/pom.xml
@@ -45,6 +45,12 @@
 
     <!-- Tests -->
     <dependency>
+      <groupId>pl.wavesoftware.sampler</groupId>
+      <artifactId>sampler-example-model</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>pl.wavesoftware.testing</groupId>
       <artifactId>junit5-starter</artifactId>
       <version>1.0.0</version>
@@ -54,12 +60,6 @@
     <dependency>
       <groupId>org.springframework</groupId>
       <artifactId>spring-test</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.projectlombok</groupId>
-      <artifactId>lombok</artifactId>
-      <version>1.18.2</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
