+++ sampler-spring/src/main/java/pl/wavesoftware/sampler/spring/SamplerScope.java
@@ -21,7 +21,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.ObjectFactory;
 import org.springframework.beans.factory.config.Scope;
-import pl.wavesoftware.sampler.core.SamplerControl;
+import pl.wavesoftware.sampler.api.SamplerControl;
 
 import java.util.Collections;
 import java.util.HashMap;
