+++ sampler-spring/src/test/java/pl/wavesoftware/sampler/example/sample/user/UserSamplerContext.java
@@ -1,0 +1,16 @@
+package pl.wavesoftware.sampler.example.sample.user;
+
+import org.springframework.context.annotation.Bean;
+import pl.wavesoftware.sampler.api.RandomSource;
+import pl.wavesoftware.sampler.api.SamplerContext;
+
+/**
+ * @author <a href="mailto:krzysztof.suszynski@wavesoftware.pl">Krzysztof Suszynski</a>
+ * @since 0.1.0
+ */
+public class UserSamplerContext {
+  @Bean
+  JohnDoe johnDoe(SamplerContext samplerContext, RandomSource randomSource) {
+    return new JohnDoeSampler(samplerContext, randomSource);
+  }
+}
