--- sampler-spring/src/test/java/pl/wavesoftware/sampler/spring/example/group/RemoteDesktop.java
@@ -1,11 +1,0 @@
-package pl.wavesoftware.sampler.spring.example.group;
-
-import pl.wavesoftware.sampler.core.Sampler;
-import pl.wavesoftware.sampler.spring.example.model.Group;
-
-/**
- * @author <a href="mailto:krzysztof.suszynski@wavesoftware.pl">Krzysztof Suszynski</a>
- * @since 0.1.0
- */
-public interface RemoteDesktop extends Sampler<Group> {
-}
