--- sampler-spring/src/test/java/pl/wavesoftware/sampler/spring/example/user/package-info.java
@@ -1,12 +1,0 @@
-/**
- * @author <a href="mailto:krzysztof.suszynski@wavesoftware.pl">Krzysztof Suszynski</a>
- * @since 0.1.0
- */
-
-@ParametersAreNonnullByDefault
-@ReturnTypesAreNonnullByDefault
-package pl.wavesoftware.sampler.spring.example.user;
-
-import pl.wavesoftware.eid.api.ReturnTypesAreNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
